moved computations to common dir
