+++ sgbatch/CMakeLists.txt
@@ -22,19 +22,25 @@ set(SOURCE_FILES
         src/SimpleExecutionController.cpp
         src/SimpleSimulator.cpp
         src/JobSpecification.h
-        src/CacheComputation.h
-        src/CacheComputation.cpp
-        src/StreamedComputation.h
-        src/StreamedComputation.cpp
-        src/CopyComputation.h
-        src/CopyComputation.cpp
+        src/computation/CacheComputation.h
+        src/computation/CacheComputation.cpp
+        src/computation/StreamedComputation.h
+        src/computation/StreamedComputation.cpp
+        src/computation/CopyComputation.h
+        src/computation/CopyComputation.cpp
         src/SimpleSimulator.h
         src/LRU_FileList.h
         )
 
 # test files
 set(TEST_FILES
-        src/JobSpecification.h src/StreamedComputation.h src/CopyComputation.h src/LRU_FileList.h src/SimpleSimulator.h)
+        src/JobSpecification.h
+        src/computation/CacheComputation.h
+        src/computation/StreamedComputation.h
+        src/computation/CopyComputation.h
+        src/LRU_FileList.h
+        src/SimpleSimulator.h
+        )
 
 # wrench library and dependencies
 find_library(WRENCH_LIBRARY NAMES wrench)
