+++ sgbatch/src/computation/StreamedComputation.cpp
