+++ sgbatch/src/computation/StreamedComputation.h
